+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.3.0 2019-05-13
+* url-encode only values, not keys of query params
+* introduce optional "stringify" callback for query-param serialization
+* enable forwarding basic-auth credentials (if exists) as default
+* Browser only: Auto support reading of "Csrf-Token" Cookie (if exits) and addint it to
+HTTP header "X-Csrf-Token" of POST, PUT and DELETE requests
+
 ## 2.0.0 2018-11-12
 * resolved, if response having http status < 300, else rejected with info object